+++ README.md
@@ -31,9 +31,7 @@ Borax 是一个Python3工具集合库。包括了以下几个话题：
 
 ## 安装 (Installation)
 
-> 从 v3.5.1开始，安装包文件格式为 *borax-3.5.1-py3-none-any.whl*（移除py2标识）以区别于之前的 *borax-3.5.0-py2.py3-none-any.whl*。
-
-Borax 要求 Python3.6+ ,可以通过 *pip* 安装 ：
+Borax 要求 Python3.5+ ,可以通过 *pip* 安装 ：
 
 ```shell
 $ pip install borax
