+++ borax/calendars/FestivalData.csv
@@ -14,15 +14,20 @@
 012250,圣诞节
 101010,春节
 101150,元宵节
+102020,龙头节
+103030,上巳节
 105050,端午节
 107070,七夕
 107150,中元节
 108150,中秋节
 109090,重阳节
 112080,腊八节
+10001A,除夕
 205026,母亲节
 206036,父亲节
 211043,感恩节
-312011,除夕
 400060,清明
 400230,冬至
+413116,初伏
+414116,中伏
+411146,末伏
