+++ borax/calendars/dataset/festivals_ext1.csv
@@ -140,7 +140,6 @@
 006230,国际奥林匹克日,public
 006230,世界手球日,public
 006250,全国土地日,public
-006260,支援酷刑受害者日,public
 006260,国际禁毒日,public
 006260,国际宪章日,public
 006300,世界青年联欢节,public
@@ -156,7 +155,6 @@
 007160,国际冰壶日,public
 007170,世界表情符号日,public
 007170,世界表情包日,public
-007180,曼德拉国际日,public
 007200,人类月球日,public
 007230,清洁工日,public
 007250,粤语日,public
@@ -176,7 +174,6 @@
 008120,国际青年节,public
 008130,国际左撇子日,public
 008150,日本无条件投降日,public
-008180,中国人力资源日,public
 008190,中国医师节,public
 008190,世界摄影日,public
 008190,世界人道主义日,public
@@ -184,19 +181,16 @@
 008260,全国律师咨询日,public
 008290,全国测绘法宣传日,public
 009010,全民健康生活方式日,public
-009010,全国中小学生开学日,public
 009030,抗日战争胜利纪念日,public
 009050,中华慈善日,public
 009080,国际扫盲日,public
 009080,国际新闻工作者日,public
 009090,全国拒绝酒驾日,public
-009090,毛泽东逝世纪念日,public
 009100,世界预防自杀日,public
 009100,教师节,basic
 009120,中国预防出生缺陷日,public
 009140,中国网民节,public
 009150,国际民主日,public
-009150,豆腐文化节,public
 009160,中国脑健康日,public
 009160,国际臭氧层保护日,public
 009170,世界骑行日,public
@@ -204,8 +198,6 @@
 009170,全民国防教育日,public
 009170,世界清洁地球日,public
 009180,全国科普日,public
-009180,九一八事变纪念日,event
-009190,国家网络安全宣传周,xz
 009200,全国爱牙日,public
 009200,公民道德宣传日,public
 009210,世界阿尔茨海默病日,public
@@ -239,7 +231,6 @@
 010100,辛亥革命纪念日,public
 010110,国际女童日,public
 010110,世界镇痛日,public
-010110,国际出柜日,public
 010120,世界关节炎日,public
 010130,中国少年先锋队诞辰日,event
 010130,国际标准时间日,public
@@ -262,7 +253,6 @@
 010240,世界发展信息日,public
 010240,联合国日,public
 010250,人类天花绝迹日,public
-010250,抗美援朝纪念日,public
 010280,世界动画日,public
 010280,世界男性健康日,public
 010310,世界城市日,public
@@ -270,13 +260,10 @@
 010310,万圣夜,public
 011010,万圣节,public
 011050,世界海啸日,public
-011070,十月社会主义革命纪念日,public
 011080,记者节,public
 011090,全国消防日,public
 011100,吉尼斯世界纪录日,public
 011100,世界青年节,public
-011110,国际科学与和平周,public
-011110,双十一购物狂欢节,public
 011110,光棍节,public
 011110,空军建军节,public
 011140,世界糖尿病日,public
@@ -318,7 +305,6 @@
 012210,国际篮球日,public
 012240,平安夜,basic
 012250,圣诞节,basic
-012260,毛泽东诞辰纪念日,life
 10001A,除夕,basic
 101010,春节,basic
 101050,破五节,tradition
@@ -352,3 +338,6 @@
 206025,文化和自然遗产日,public
 207015,国际合作节,public
 209025,世界急救日,public
+413116,初伏,tradition
+414116,中伏,tradition
+411146,末伏,tradition
