+++ docs/guides/festivals2.md
@@ -66,7 +66,7 @@ ld = wd.lunar
 print(sd) # 2021-01-01
 print(ld) # LunarDate(2020, 11, 18, 0)
 ```
- ### 日期计算
+### 日期计算
 
 `WrappedDate` 也支持日期的计算。原则是 **只要操作数中有一个是WrappedDate对象，结果就是WrappedDate对象**。
 
@@ -79,7 +79,7 @@ print(ld) # LunarDate(2020, 11, 18, 0)
 | WappedDate       | -      | timedelta        | WappedDate  |
 
 
- ### 序列化
+### 序列化
 
  `WrappedDate` 支持编码序列化，参见 日期序列化 一节。
 
@@ -245,8 +245,6 @@ TermFestival(index=0)
 
 类型：str，节日的标准化描述，如“公历每年1月1日”、“公历每年八月十五”、“公历每年6月第2个星期六”等。
 
-设置节日对象的 name。
-
 例子
 
 ```python
@@ -260,7 +258,11 @@ print(LunarFestival(month=1, day=1).description) # '农历每年正月初一'
 print(WeekFestival(month=5, index=2, week=calendar.SUNDAY, name='母亲节').description) # '公历5月第2个星期日'
 ```
 
+### catalog
 
+> Add in 3.5.6
+
+类型:str，节日的分类标识。
 
 ## Festival API
 
@@ -451,6 +453,40 @@ FestivalLibrary.get_festival_names(self, date_obj: MixedDate) -> list
 
 获取某一个日期的节日名称列表。
 
+### list_days_in_countdown
+
+> Add in 3.5.6
+
+```python
+FestivalLibrary.list_days_in_countdown(countdown: Optional[int] = None, date_obj: MixedDate = None
+    ) -> List[Tuple[int, WrappedDate, Festival]]
+```
+
+迭代获取某个时间的倒计时信息。
+
+计算节日及其距离今天（2021年5月4日）的天数
+
+```python
+
+from borax.calendars.festivals2 import FestivalLibrary
+
+library = FestivalLibrary.load_builtin()
+for ndays, wd, festival in library.list_days_in_countdown(countdown=365):
+    print(f'{ndays:>3d} {wd.simple_str()} {festival.name}')
+```
+
+输出结果
+
+```
+  0 2022-05-04(四月初四) 青年节
+  4 2022-05-08(四月初八) 母亲节
+  8 2022-05-12(四月十二) 护士节
+...
+332 2023-04-01(闰二月十一) 愚人节
+336 2023-04-05(闰二月十五) 清明
+362 2023-05-01(三月十二) 劳动节
+```
+
 ### iter_festival_countdown
 
 ```python
@@ -476,28 +512,7 @@ for nday, gd_list in fl.iter_festival_countdown():
   7 儿童节 2021-06-01(二〇二一年四月廿一)
  20 端午节 2021-06-14(二〇二一年五月初五)
  26 父亲节 2021-06-20(二〇二一年五月十一)
- 68 建军节 2021-08-01(二〇二一年六月廿三)
- 81 七夕 2021-08-14(二〇二一年七月初七)
- 89 中元节 2021-08-22(二〇二一年七月十五)
-108 教师节 2021-09-10(二〇二一年八月初四)
-119 中秋节 2021-09-21(二〇二一年八月十五)
-129 国庆节 2021-10-01(二〇二一年八月廿五)
-142 重阳节 2021-10-14(二〇二一年九月初九)
-184 感恩节 2021-11-25(二〇二一年十月廿一)
-210 冬至 2021-12-21(二〇二一年冬月十八)
-213 平安夜 2021-12-24(二〇二一年冬月廿一)
-214 圣诞节 2021-12-25(二〇二一年冬月廿二)
-221 元旦 2022-01-01(二〇二一年冬月廿九)
-230 腊八节 2022-01-10(二〇二一年腊月初八)
-251 除夕 2022-01-31(二〇二一年腊月廿九)
-252 春节 2022-02-01(二〇二二年正月初一)
-265 情人节 2022-02-14(二〇二二年正月十四)
-266 元宵节 2022-02-15(二〇二二年正月十五)
-287 妇女节 2022-03-08(二〇二二年二月初六)
-291 植树节 2022-03-12(二〇二二年二月初十)
-311 愚人节 2022-04-01(二〇二二年三月初一)
-315 清明 2022-04-05(二〇二二年三月初五)
-341 劳动节 2022-05-01(二〇二二年四月初一)
+...
 344 青年节 2022-05-04(二〇二二年四月初四)
 348 母亲节 2022-05-08(二〇二二年四月初八)
 352 护士节 2022-05-12(二〇二二年四月十二)
@@ -596,52 +611,8 @@ FestivalLibrary.to_csv(path_or_buf)
 保存到 csv 文件。
 
 
-## 综合使用示例
-
-###  两头春、无头春
-
-```python
-# 农历两头春，无头春。在农历year年时间段，立春的个数。
-
-from borax.calendars.festivals2 import TermFestival, Period
-from borax.calendars.lunardate import MIN_LUNAR_YEAR, MAX_LUNAR_YEAR
-
-tf = TermFestival(name='立春')
-for year in range(MIN_LUNAR_YEAR, MAX_LUNAR_YEAR):
-    start_date, end_date = Period.lunar_year(year)
-    ncount = len(tf.list_days(start_date, end_date))
-    print('{}({}) - {}({})    {}'.format(
-        start_date.cn_str(),
-        start_date.to_solar_date(),
-        end_date.cn_str(),
-        end_date.to_solar_date(),
-        ncount
-    ))
-
-```
 
 
-输出
-```
-...
- 
-二〇一一年正月初一(2011-02-03) - 二〇一一年腊月廿九(2012-01-22)    1
-二〇一二年正月初一(2012-01-23) - 二〇一二年腊月廿九(2013-02-09)    2
-二〇一三年正月初一(2013-02-10) - 二〇一三年腊月三十(2014-01-30)    0
-二〇一四年正月初一(2014-01-31) - 二〇一四年腊月三十(2015-02-18)    2
-二〇一五年正月初一(2015-02-19) - 二〇一五年腊月廿九(2016-02-07)    1
-二〇一六年正月初一(2016-02-08) - 二〇一六年腊月三十(2017-01-27)    0
-二〇一七年正月初一(2017-01-28) - 二〇一七年腊月三十(2018-02-15)    2
-二〇一八年正月初一(2018-02-16) - 二〇一八年腊月三十(2019-02-04)    1
-二〇一九年正月初一(2019-02-05) - 二〇一九年腊月三十(2020-01-24)    0
-二〇二〇年正月初一(2020-01-25) - 二〇二〇年腊月三十(2021-02-11)    2
-二〇二一年正月初一(2021-02-12) - 二〇二一年腊月廿九(2022-01-31)    0
-二〇二二年正月初一(2022-02-01) - 二〇二二年腊月三十(2023-01-21)    1
-二〇二三年正月初一(2023-01-22) - 二〇二三年腊月三十(2024-02-09)    2
-二〇二四年正月初一(2024-02-10) - 二〇二四年腊月廿九(2025-01-28)    0
-二〇二五年正月初一(2025-01-29) - 二〇二五年腊月廿九(2026-02-16)    2
-二〇二六年正月初一(2026-02-17) - 二〇二六年腊月廿九(2027-02-05)    1
- 
-...
-```
+
+
