+++ docs/guides/fetch.md
@@ -74,6 +74,45 @@ print(ages)
 [30, 56, 56]
 ```
 
+### 列表型数据
+
+`fetch` 函数第一个参数 `iterable` 也可以是列表型数据。
+
+```python
+from borax.datasets.fetch import fetch
+
+data = [
+    [0, 1, 2, 3, 4, 5, 6, 7, 8, 9],
+    [10, 11, 12, 13, 14, 15, 16, 17, 18, 19],
+    [20, 21, 22, 23, 24, 25, 26, 27, 28, 29]
+]
+zeros, twos = fetch(data, 0, 2)
+print(zeros)
+print(twos)
+```
+
+输出
+
+```
+[0, 10, 20]
+[2, 12, 22]
+```
+
+此时情况下和 `zip` 函数有相类似的功能，上述例子可改写如下：
+
+```python
+data = [
+    [0, 1, 2, 3, 4, 5, 6, 7, 8, 9],
+    [10, 11, 12, 13, 14, 15, 16, 17, 18, 19],
+    [20, 21, 22, 23, 24, 25, 26, 27, 28, 29]
+]
+zeros, _, twos, *_ = zip(*data)
+print(zeros)
+print(twos)
+```
+
+
+
 ## 提供默认值
 
 当 `iterable` 数据列表缺少某个属性/键，可以通过指定 `default` 或 `defaults` 参数提供默认值。
