+++ mkdocs.yaml
@@ -20,28 +20,27 @@ markdown_extensions:
   - pymdownx.superfences
 nav:
   - 首页: README.md
-  - 中国农历:
+  - 农历与节日:
     - 农历: guides/lunardate.md
     - 节日: guides/festivals2.md
-    - 日期节日序列化: guides/festivals2-serialize.md
+    - 序列化: guides/festivals2-serialize.md
     - 生日: guides/birthday.md
     - 工具类: guides/calendars-utils.md
-  - 数据结构:
+  - 数据结构与数字:
     - 数据连接(Join): guides/join.md
     - 列选择器(fetch): guides/fetch.md
     - 树形结构: guides/tree.md
     - cjson: guides/cjson.md
-  - 数字:
     - 中文数字: guides/numbers.md
     - 百分数: guides/percentage.md
-  - 其他:
+  - 其他模块:
     - 字符串: guides/strings.md
     - 单例模式: guides/singleton.md
     - 序列号池: guides/serial_pool.md
-    - json: guides/cjson.md
   - 文章:
     - 农历与节日: guides/festivals2-usage.md
-  - 项目日志:
+    - 农历开发笔记: posts/lunardate-development.md
+  - 更新日志:
     - 更新日志: changelog.md
     - 发布日志:
       - v3.5.6: release-note/v356.md
