+++ setup.cfg
@@ -1,6 +1,10 @@
 [metadata]
 long_description = file: README.md
 long_description_content_type = text/markdown
+project_urls =
+    Homepage = https://borax.readthedocs.io/zh_CN/latest/
+    Source Code = https://github.com/kinegratii/borax
+    Bug Tracker = https://github.com/kinegratii/borax/issues
 
 [nosetests]
 traverse-namespace = 1
