+++ setup.py
@@ -34,7 +34,6 @@ setup(
     python_requires='>=3.5',
     packages=find_packages(exclude=['tests']),
     include_package_data=True,
-    url='https://github.com/kinegratii/borax',
     license='MIT',
     author='kinegratii',
     author_email='zhenwei.yan@hotmail.com',
